Retain and report the most recent status of a feed update.
If a feed fails to update, the reason will be stored in the newly-added column 'feed_cache_last_attempt_status' in FOF_FEED_TABLE.  The new function fof_db_update_attempt_status() will set and clear the value.

Fixed some of the SimplePie error-checking to use the public error method rather than the internal error property.

Irregular statuses are now indicated in the sidebar feed icon slot, as well as on the feed detail page.