Reimplemented 'delete tag from all items' action to not use all of the memories.
Added fof_db_untag_user_all() to fof-db.php, which simply removes all occurrences of a tag for a user.
Changed fof_untag() in fof-main.php to use this function instead of generating a list of tagged items and then removing that tag from those items.  The old method worked for a handful of items, but was using a ridiculously unnecessary amount of memory when trying to clear a tag from a few tens-of-thousands of items.