Let feed items get updated
If an item is updated in the feed, we already do pretty much all the work necessary to update its content, title, etc.; why not just let it all get updated? This also rolls the fof_db_add_item and fof_db_update_item into a single function. Because why not.