A whole bunch of changes to support text entries. For example:
* Removed athe spurious 'balancetags' plugin (which is now more than covered by other filters and by SimplePie itself)
* Allowed div, pre, and span into the plain plugin (since problematic uses of those tags are already handled by SimplePie)
* text content now properly detects text and unspecified content type on Atom feeds, and wraps the items in an appropriate div (which now has a stylesheet entry)

This also means there's no need for a custom sanitizer (which is good, because it never really worked right in the first place) and so on.

Also, found an annoying issue where the image proxy was running all the render filters even though it had no need to, so I allow for something to just get an item without doing that. (I suppose I could have also just made img use the fof_db_get_item version instead.)