Explicitly mark an item as complete
This adds a new column, `item_complete`, which marks that all tags and plugin settings have been successfully applied to a new item. This is an attempt at fixing #25 without really understanding the underlying cause.

Unfortunately transactions could not be used for this because of how concurrent transactions (don't) work in SQLite and PDO.